added posibility for powers and made floats

+++ cpp09/ex01/include/RPN.hpp
@@ -5,7 +5,7 @@
 class RPN
 {
 private:
-	std::stack< int > _stack;
+	std::stack< float > _stack;
 
 public:
 	RPN();
